MISC: Merge

+++ src/lib/itkTextureFilter/itkScalarImageToIntensitySizeListSampleFilter.h
@@ -94,17 +94,17 @@ public:
     itkSetMacro(NumberOfIntensityBins, int)
     itkGetMacro(NumberOfIntensityBins, int)
 
-    itkSetMacro(MinSize, int)
-    itkGetMacro(MinSize, int)
+    itkSetMacro(MinSize, double)
+    itkGetMacro(MinSize, double)
 
-    itkSetMacro(MaxSize, int)
-    itkGetMacro(MaxSize, int)
+    itkSetMacro(MaxSize, double)
+    itkGetMacro(MaxSize, double)
 
     itkSetMacro(BackgroundValue, InputImagePixelType)
     itkGetMacro(BackgroundValue, InputImagePixelType)
 
-    itkSetMacro(IntensityBinSize, int)
-    itkGetMacro(IntensityBinSize, int)
+    itkSetMacro(IntensityBinSize, double)
+    itkGetMacro(IntensityBinSize, double)
 
 protected:
     ScalarImageToIntensitySizeListSampleFilter();
@@ -127,9 +127,9 @@ private:
   InputImagePixelType m_BackgroundValue;
   int m_NumberOfIntensityBins;
 
-  int m_MinSize;
-  int m_MaxSize;
-  int m_IntensityBinSize;
+  double m_MinSize;
+  double m_MaxSize;
+  double m_IntensityBinSize;
 
 };
